Adding decline and reset button (#13)
## Why do we need this?

We want reviewers to be able to decline themselves and others if they are on vacation or unnecessary on the review. If you get declined, you also want to be able to get reset (aka. re-added).

## What does this change do?

Add decline and reset buttons. These post specially formatted comments on the review request that we later parse to add status UI for these users.

![image](https://user-images.githubusercontent.com/3143819/83913290-eb894000-a734-11ea-93dc-b32df2434724.png)

Declined users stay on the review request since only the owner of the request can remove them (because of permissions). They also get emails on declines and resets -- a good thing; we want to let our reviewers know when this happens to them so they can react if needed.

## How was this tested?

Manual testing on Chrome and FireFox.

## Checklist

If applicable:

- [x] Runs in latest stable Chrome and Firefox without Javascript errors
- [x] Does not report any issues with `eslint` (run `npm run build` -- the PR build will also check this)
- [x] Increment version following [semver](https://semver.org/)
- [x] All `@require` and `@resource` URLs must have SRI ([use this hashing tool](https://www.srihash.org/))
- [x] Blocks of code are commented
- [x] Included a screenshot in the PR
- [x] Added docs to `README.md`
- [ ] Linked issues